Fix ARM64 warnings in EventManager, ProcessMgr, Dialog, and more
- Fix pointer-to-int-cast warnings in SystemEvents.c, ProcessManager.c
- Fix int-to-pointer-cast warnings in LowMemGlobals.c
- Remove unreachable type-limits comparisons in DialogManagerCore.c
- Remove unreachable type-limits comparisons in DialogDrawing.c
- Remove unreachable type-limits comparisons in StringUtilities.c

+++ src/CPU/m68k_interp/LowMemGlobals.c
@@ -193,10 +193,10 @@ void LMSetApplZone(UInt32 value)
 
 void* LMGetThePort(void)
 {
-    return (void*)LMGetLong(LMG_ThePort);
+    return (void*)(uintptr_t)LMGetLong(LMG_ThePort);
 }
 
 void LMSetThePort(void* port)
 {
-    LMSetLong(LMG_ThePort, (UInt32)port);
+    LMSetLong(LMG_ThePort, (UInt32)(uintptr_t)port);
 }
